pyproject.toml
[tool.poetry]
name = "task-management-restfull-api"
version = "0.1.0"
description = "🐍 A RESTful API for a Task Management app built with Python, Flask, and MongoDB. Manage projects, tasks, and subtasks with features like authentication, caching, pagination, and more! Follows OOP principles, SOLID, and best practices. Designed to be fast, scalable, and developer-friendly!"
authors = ["Ismael Pamplona"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
pymongo = "^4.9.1"
python-dotenv = "^1.0.1"
bcrypt = "^4.2.0"
flask-jwt-extended = "^4.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
pytest-cov = "^5.0.0"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
